Fixed, but not
After the encoders broke (because its the day before meet 0 of course), we figured out the Arm encoder messed up the Left and Right motor encoders. Unplugging it, it fixed the issue of the Left and Right encoders not working, but in turn messed up all of our A-D programs.