Fixed Encoders and Teleop
The encoders were fixed by correcting the motor AND encoder cables and config since they both had an error. For the teleop, it was using the autonomous' hardware map instead of its own unique one, so we defined the motors and such in the teleop program (like we always have).