Changed teleops and added Zeus + encoder drives works now
The thing is, for some reason the Vuforia in the aggregated code is not working. To fix, compare it to the code that merely checks the camera output. From testing, I deduced it is the actual code, not the way I call to it that is making the camera not turn on properly. 

TL;DR
Something is different that is making the init_vuforia and start_vuforia methods not turn the camera on.