Finished Hermes D and D_2 Started Hermes A
Finished all three positions for Hermes A and A_2. A_2 needs further testing and adjustment. Hermes A needs to be fixed when it comes to detecting positions with swapped Color Sensors (since Color Sensor 1 is reading position 2 rather than position 1). There is also an error in Hermes A where the second movement, the strafe to the left, would go significantly longer than it is programmed to do.