Finished A Main, Fast, Foundation, and Started D Max
A Main, Fast, and Foundation is done and ready to be used. D Main, Fast and Foundation is also done and ready to be used. D Max was started and position 1 is ready to be used. Position 1 still needs to be fine tuned and there is a error where the robot can not find position 2 and will run position 1. We will need 2 Max programs, 1 that parks and 1 that doesn't,