Made Major Tuning to all Codes
Hermes Maxes is put on hold unto Championships
Hermes Foundation, Fast, and Default was further tuned. Final test on Wednesday 2/5. Hermes Place was created. This code places 1 SkyStone and leaves another outside of the Foundation. This program is used to get the 2 SkyStone points while also allowing the Alliance team to bring the Foundation to the Build Site. Further tuning on Hermes C and Bs were made. 
TODO:
For max, we can use Foundation and speed it up just a little since getting 1 SkyStone for Foundation only needs 15 seconds to place. If we speed little parts up we can cut this speed to around 10 seconds, allowing 20 seconds for SkyStone placement and 10 seconds for re positioning of Foundation. This would be more than enough time. However, a major issue we have with speeding up SkyStone placement would be the extreme accuracy loss. Because of this, this code was delayed to the Championship (If we make it). A good solution for this problem is most likely encoders with PID. This would allow us to drive at full speed while avoiding the drift issues.